Deduplicate settings and processes
